Revert "Feature/pypackage"
