--- .gitignore
@@ -1,132 +1,0 @@
-# General
-.DS_Store
-.idea
-.ipynb_checkpoints
-*.pyc
-*~
-.envrc
-.direnv
-.commit
-
-rp-indexer
-
-# temba
-dump.rdb
-.coverage
-coverage_html_report
-env
-build
-config/gen
-config/settings.py
-coverage-report
-celerybeat-schedule.db
-db
-node_modules
-parser.out
-screenshots
-sitestatic
-static/coffee/analytics.js
-static/coffee/flows.js
-static/coffee/simulator.js
-static/coffee/temba.js
-static/coffee/formax.js
-static/javascripts/
-temba/settings.py
-temba/smartmin
-temba.db
-smartmin
-mailroom_test.dump
-docs/
-
-# Exclude media folder except files used for import testing
-media/apks
-media/contacts
-media/csv_imports
-media/flow
-media/messages
-media/orgs
-media/recordings
-media/test_orgs
-media/tmp
-media/attachments
-
-
-# External projects
-fabric
-fabconfig.py
-fabfile.py
-deploy
-config
-Flows.ipynb
-
-# Byte-compiled / optimized / DLL files
-__pycache__/
-*.py[cod]
-
-# C extensions
-*.so
-
-# Distribution / packaging
-.Python
-env/
-venv/
-build/
-develop-eggs/
-dist/
-downloads/
-eggs/
-lib/
-lib64/
-parts/
-sdist/
-var/
-*.egg-info/
-.installed.cfg
-*.egg
-
-# PyInstaller
-#  Usually these files are written by a python script from a template
-#  before PyInstaller builds the exe, so as to inject date/other infos into it.
-*.manifest
-*.spec
-
-# Installer logs
-pip-log.txt
-pip-delete-this-directory.txt
-
-# Unit test / coverage reports
-htmlcov/
-.tox/
-.coverage
-.cache
-.perf_results
-nosetests.xml
-coverage.xml
-js-coverage/
-
-# Translations
-*.pot
-
-# Django stuff:
-*.log
-
-# Sphinx documentation
-docs/_build/
-
-# PyBuilder
-target/
-textit
-
-# Vagrant/Ansible
-.vagrant/
-ansible/vendor/
-ansible/site.retry
-
-# sublime projects
-*.sublime*
-
-# VS Code projects
-.vscode/
-
-# autoenv
-.direnv/
