--- README.md
@@ -1,19 +1,0 @@
-# RapidPro Apps
-
-## Installing
-```shell script
-$ pip install -e git+git://github.com/Ilhasoft/rapidpro-apps.git
-```
-
-Add the desired apps to you INSTALLED_APPS setting:
-
-**settings.py**
-```python
-...
-
-INSTALLED_APPS += ("weni.<app_name>", ...)
-```
-## APPs available
-
-- [Channel Stats](channel_stats/README.md)
-- [Analytics API](analytics_api/README.md)
