--- setup.cfg
@@ -1,26 +1,0 @@
-[metadata]
-name = rapidpro-apps
-version = 0.1
-description = Weni apps for Rapidpro Platform
-long_description = file: README.md
-url = https://github.com/Ilhasoft/rapidpro-apps
-author = jcbalmeida
-author_email = jcbalmeida@gmail.com
-licence = AGPL-3.0
-classifiers =
-    Environment :: Web Environment
-    Framework :: Django
-    Framework :: Django :: 2.1
-    Intended Audience :: Developers
-    License :: OSI Approved :: GNU Affero General Public License v3
-    Operating System :: OS Independent
-    Programming Language :: Python
-    Programming Language :: Python :: 3
-    Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.6
-    Topic :: Internet :: WWW/HTTP
-    Topic :: Internet :: WWW/HTTP :: Dynamic Content
-
-[options]
-include_package_data = true
-packages = find:
