--- weni/analytics_api/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = "weni.analytics_api.apps.AnalyticsApiConfig"
