--- weni/analytics_api/apps.py
@@ -1,11 +1,0 @@
-from django.apps import AppConfig
-
-
-class AnalyticsApiConfig(AppConfig):
-    name = "weni.analytics_api"
-
-    def ready(self):
-        from .urls import urlpatterns
-        from ..utils.app_config import update_urlpatterns
-
-        update_urlpatterns(urlpatterns)
