--- weni/channel_stats/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = "weni.channel_stats.apps.ChannelStatsConfig"
